requirements/base.txt
Django>=3.0.7,<3.1.0
djangorestframework>=3.11.0,<3.12.0
requests>=2.24.0,<2.25.0
django-rest-knox==4.1.0
